refactor: 💡 separate pkg
